+++ xgboostspark/src/main/scala/org/dmlc/tracker/TrackerMessages.scala
@@ -1,0 +1,7 @@
+package org.dmlc.tracker
+
+//messages from TaskProxy to JobTracker
+final case class RegisterTaskProxy(workerId: Int)
+
+//messages from JobTracker to TaskProxy
+case object StartTraining
